+++ Formatter/IValueFormatter.cs
@@ -4,6 +4,6 @@ namespace XF.ChartLibrary.Formatter
 {
     public interface IValueFormatter
     {
-        string GetFormattedValue(double value, Data.Entry entry, int dataSetIndex, ViewPortHandler viewPortHandler);
+        string GetFormattedValue(float value, Data.Entry entry, int dataSetIndex, ViewPortHandler viewPortHandler);
     }
 }
