+++ Platform/Droid/Data/EntryBase.cs
@@ -4,12 +4,12 @@ namespace XF.ChartLibrary.Data
 {
     public partial class EntryBase
     {
-        public EntryBase(double y, Drawable icon) : this(y)
+        public EntryBase(float y, Drawable icon) : this(y)
         {
             Icon = icon;
         }
 
-        public EntryBase(double y, Drawable icon, object data) : this(y)
+        public EntryBase(float y, Drawable icon, object data) : this(y)
         {
             Icon = icon;
             Data = data;
