+++ Platform/NetStandard/Data/EntryBase.cs
@@ -4,12 +4,12 @@ namespace XF.ChartLibrary.Data
 {
     public partial class EntryBase
     {
-        public EntryBase(double y, SKImage icon) : this(y)
+        public EntryBase(float y, SKImage icon) : this(y)
         {
             Icon = icon;
         }
 
-        public EntryBase(double y, SKImage icon, object data) : this(y)
+        public EntryBase(float y, SKImage icon, object data) : this(y)
         {
             Icon = icon;
             Data = data;
